requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
psycopg2-binary==2.9.9
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
python-decouple==3.8
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Code Quality Tools
black==23.12.1
flake8==7.0.0
isort==5.13.2
djlint==1.34.0
